pyproject.toml
[project]
name = "spendscan"
version = "0.1.0"
description = "Aplikacja analizująca wydatki na podstawie zdjęć paragonów"
readme = "README.md"
requires-python = ">=3.13"
license = { text = "MIT" }
authors = [
    { name = "Piotr Marczak" },
    { name = "Jakub Bryła" },
    { name = "Mateusz Mróz" },
    { name = "Igor Typiński" },
    { name = "Mateusz Słoń" },
]
dependencies = [
    "fastapi>=0.136.1",
    "google-genai>=1.73.1",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "numpy>=2.4.4",
    "pillow>=12.2.0",
    "platformdirs>=4.9.6",
    "pydantic>=2.13.3",
    "pydantic-settings>=2.14.0",
    "python-multipart>=0.0.26",
    "tenacity>=9.1.4",
]

[dependency-groups]
dev = [
    { include-group = "lint" },
    { include-group = "test" },
    { include-group = "type" },
    "pre-commit>=4.5.1",
]
lint = ["ruff>=0.11"]
test = ["pytest>=8", "pytest-cov>=6", "coverage[toml]>=7"]
type = ["mypy>=1.15"]

[tool.uv]
default-groups = ["dev"]

# ── ruff ──────────────────────────────────────────────

[tool.ruff]
target-version = "py313"
line-length = 120
indent-width = 4

[tool.ruff.lint]
select = [
    # 🔴 Tier 1: MUSISZ
    "F", "E", "W", "I", "N", "UP", "B", "S", "BLE", "G", "T20", "PGH", "RUF",
    # 🟡 Tier 2: POWINIENEŚ
    "C4", "SIM", "RET", "PTH", "PT", "TRY", "PERF", "PL", "LOG",
]
ignore = [
    "PLR0913",  # Too many arguments — normalne w config functions
    "PLR2004",  # Magic value comparison — `if status == 200:` jest czytelne
    "TRY003",   # Long message in exception — overengineering
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101"]
"scripts/**" = ["T20"]
"__init__.py" = ["F401"]

[tool.ruff.lint.isort]
known-first-party = ["spendscan"]

[tool.ruff.lint.pylint]
max-args = 8

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true

# ── mypy ──────────────────────────────────────────────

[tool.mypy]
strict = true
mypy_path = "backend"
warn_unreachable = true
pretty = true
show_error_codes = true
show_error_code_links = true
python_version = "3.13"
enable_error_code = ["deprecated"]

[[tool.mypy.overrides]]
module = ["tests.*"]
disallow_untyped_defs = false

# ── pytest ────────────────────────────────────────────

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["backend"]
addopts = ["-ra", "--strict-markers", "--strict-config"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks integration tests",
]

# ── coverage ──────────────────────────────────────────

[tool.coverage.run]
source = ["spendscan"]
branch = true

[tool.coverage.report]
show_missing = true
fail_under = 50
